created boss with bullets, boss level at 10, clears enemies then drops boss to attack for the rest of the wave, gradually getting faster till boss reaches its max speed

+++ src/com.tutorial.main/Game.java
@@ -26,6 +26,7 @@ public class Game extends Canvas implements Runnable {
         handler.addObj(new Player(WIDTH / 2 - 32, HEIGHT / 2 - 32, ID.Player, handler));
 //        for(int i=0; i<4; i++) {
             handler.addObj(new BasicEnemy(r.nextInt(WIDTH-20), r.nextInt(HEIGHT-20), ID.BasicEnemy, handler));
+
 //        }
 //        handler.addObj(new Player(WIDTH/2+64, HEIGHT/2-32, ID.Player2));
 //        handler.addObj(new Player(100,100,ID.Player));
@@ -104,7 +105,7 @@ public class Game extends Canvas implements Runnable {
         bs.show();
     }
 
-    public static int clamp(int var, int min, int max) {
+    public static float clamp(float var, float min, float max) {
         if (var >= max) {
             return var = max;
         } else if (var <= min) {
